+++ ProyectosSDL/HolaSDL/Vector2D.h
@@ -1,26 +1,27 @@
 #pragma once
+#include "checkML.h"
 
 using namespace std;
 //using Point2D = Vector2D;
 
 class Vector2D {
 private:
-	int X = 0;
-	int Y = 0;
+	double X = 0;
+	double Y = 0;
 
 public:
 	Vector2D() {};
-	Vector2D(int x, int y) { X = x; Y = y; };
+	Vector2D(double x, double y) { X = x; Y = y; };
 
-	int getX() const { return X; };
-	int getY() const { return Y; };
-	void setX(int valor) { X = valor; };
-	void setY(int valor) { Y = valor; };	
+	double getX() const { return X; };
+	double getY() const { return Y; };
+	void setX(double valor) { X = valor; };
+	void setY(double valor) { Y = valor; };
 	void subVectorInY(Vector2D vel) { Y -= vel.Y; };
-	void multiply(int n) { X* n; Y* n; };
+	void multiply(double n) { X* n; Y* n; };
 	void normalize();
 	Vector2D operator+(const Vector2D& v) const;
-	Vector2D operator*(int d) const;
-	int operator*(const Vector2D& d) const;
+	Vector2D operator*(double d) const;
+	double operator*(const Vector2D& d) const;
 };
